Fixed a couple AI bugs
For Protoman.

The panels flashed twice in one attack, which also looked a bit weird.

Also, whenever the player was in front of protoman before he was meant to attack, the player took damage.

To fix those two bugs, I redid the ai code to execute based on events, instead of timers. Will likely overhaul the enemy AI system to work more on events than timers at some point in the future.

On another note, I applied the panel-flashing to the Hammer AI.