Added Game Over, Stage Select, and Victory functionality
When any combatant dies, their Died event invokes, disabling their game objects (best not to destroy them). When all enemies are dead, the victory sequence executes. When you die, the Game Over sequence executes.

Much of this is thanks to the EndOfBattleController, which lets you set which scene loads when you die, and which scene loads when you win.